Add memory mappings and TI CF62011BPC emulation
This is a über commit that adds a lot of stuff.
First part: complete memory subsystem rewrite with support for memory
mappings and 32-bit CPU bus emulation.
Second part: emulation of the TI CF62011BPC video chip. Bare bones, with
the only purpose of passing the 2121's POST procedure. The CF62011BPC
chip seems to be an XGA clone, for which there's plenty of
documentation, so maybe VESA emulation could be possible (at least in
theory).

WARNING: code not thoroughly tested.