Add Sound Blaster emulation
This is a bare bones implementation.
Not thoroughly tested and only on a handful of games and demos.

Models emulated: Sound Blaster 1.5 and 2.0 (no CMS chips, OPL2 only).

There's already some Pro models code, but the mixer chip is missing so
they are not functioning. Sound Blaster 16 emulation is still far away.