Resample audio when Machine's speed is not 100%
Stretch or shrink audio data produced by emulated audio cards to what's
needed by the output sound device, when the emulation speed is not 100%.
This allows for continuous sound output even when the Machine is
accelerated or slowed, either intentionally or when the load is
excessive.
Audio is resampled only for the output sound device, when capturing the
result will always be at 100% speed.