Fix filesystem access on Windows
Closes #43 (hopefully, for now at least).

As it turns out, strings on Windows are a shitshow (surprise). They have
2 different possible encodings, both insane. While the entire civilized
world settled on UTF-8 (the only true encoding), Microsoft decided to
play a different game and now Win32 has ANSI and UTF-16, with a vast
amount of string-related functions with 2 possible versions (A and W),
selected by fucking macros. And on top of that delirium, filesystem
paths add another level of insanity.