Fix a tmp file related launch error on Windows
Remove POSIX tmpfile() and use a (stinky) custom solution, because MinGW
implementation of tmpfile() doesn't work. It apparently creates a
temporary file in the root directory of the drive from which IBMulator
is launched. If the drive is C: then an error occurs because the user
isn't usually allowed to write there. It's either that or some other
permission related issue...