Reimplement HFE floppy disk image format
Based on MAME's 2d5bca1 and 4adf0ce, with the following differences:

 * when reading the bitstream, wrap around instead of reloading the
previous byte (cell group) when running out of samples
 * when saving the bitstream, fill the entire track buffer, wrapping
around instead of 0-fill when reaching the end of the track

The curent implementation can't handle correctly the case of slightly
shorter than expected tracks. As a workaround it is assumed that both
the beginning and end of track are GAP bytes.